+++ CHANGELOG.md
@@ -15,6 +15,7 @@ develop
 ### Fixes
 
 - Fix an issue where the selected marks of line height / tab width in the Format menu disappeared.
+- Fix an issue where unselected last line number could be highlighted if the text orientation is vertical.